app/core/security.py
from datetime import datetime, timedelta, timezone
from typing import Any

import jwt
from passlib.hash import pbkdf2_sha256

from app.core.config import settings


def create_access_token(sub: Any, expires_delta: timedelta | None = None) -> str:
    expire = datetime.now(timezone.utc) + (
        expires_delta if expires_delta else timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    )
    to_encode = {"exp": expire, "sub": sub}
    encoded_jwt = jwt.encode(to_encode, algorithm=settings.JWT_ALGORITHM, key=settings.JWT_SECRET_KEY)
    return encoded_jwt


def decode_access_token(token: str) -> Any:
    return jwt.decode(str(token), key=settings.JWT_SECRET_KEY, algorithms=[settings.JWT_ALGORITHM])


def get_password_hash(password: str) -> str:
    return pbkdf2_sha256.hash(password)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    return pbkdf2_sha256.verify(plain_password, hashed_password)
