pyproject.toml
[project]
name = "authorization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "grpcio>=1.71.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.71.0",
    "ruff>=0.11.11",
]
